Fix type in struct ulp_detour
The universe counter has unsigned long type across the code base, except
in struct ulp_detour. This patch fixes the mismatch. Even though this
counts as an ABI break, libpulp has never been released, so the SONAME
is kept the same.

+++ include/ulp.h
@@ -54,7 +54,7 @@ struct ulp_detour_root {
 
 struct ulp_detour {
     unsigned char patch_id[32];
-    unsigned int universe;
+    unsigned long universe;
     void *target_addr;
     char active;
     struct ulp_detour *next;
